Shuffle around the structure of the project
- move most code into train_bot_uk/train_bot_uk.py
- move out responder code into train_bot_uk/responders/
- make a tests/ directory
