+++ train_bot_uk/__init__.py
@@ -1,0 +1,1 @@
+from train_bot_uk import *
