+++ train_bot_uk/responders/__init__.py
@@ -1,0 +1,7 @@
+#!/usr/bin/env python
+# encoding: utf-8
+
+from __future__ import unicode_literals
+
+from journey_responder import JourneyResponder
+from ping_responder import PingResponder
